This is my database homework about spatial database
In this homework, you are going to work with spatial data - you will
create some data, visualize it, do queries on it, and visualize the
query results.. Hope you have fun with this!

The exercise will give you a taste of various spatial data-related
programs and file formats and query functions, all of which are quite
useful from a real-world (or job interview) perspective.